+++ packages/slate-sync/index.js
@@ -2,8 +2,8 @@ const chalk = require('chalk');
 const figures = require('figures');
 const https = require('https');
 const themekit = require('@shopify/themekit').command;
-const slateEnv = require('@shopify/slate-env');
-const SlateConfig = require('@shopify/slate-config');
+const slateEnv = require('@by-association-only/slate-env');
+const SlateConfig = require('@by-association-only/slate-config');
 
 const config = new SlateConfig(require('./slate-sync.schema'));