+++ packages/slate-tools/tools/theme-lint/index.js
@@ -1,5 +1,5 @@
 const execSync = require('child_process').execSync;
-const SlateConfig = require('@shopify/slate-config');
+const SlateConfig = require('@by-association-only/slate-config');
 
 const config = new SlateConfig(require('../../slate-tools.schema'));