+++ packages/slate-tools/tools/webpack/config/parts/css.js
@@ -1,5 +1,5 @@
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const SlateConfig = require('@shopify/slate-config');
+const SlateConfig = require('@by-association-only/slate-config');
 const config = new SlateConfig(require('../../../../slate-tools.schema'));
 
 const isDev = process.env.NODE_ENV === 'development';
@@ -40,7 +40,7 @@ const postcssLoader = {
   },
 };
 
-const cssVarLoader = {loader: '@shopify/slate-cssvar-loader'};
+const cssVarLoader = {loader: '@by-association-only/slate-cssvar-loader'};
 
 cssRule.use = [
   ...(isDev ? [styleLoader] : [MiniCssExtractPlugin.loader, cssVarLoader]),